Move `PatientSummaryProfile` to the `.summary` package
